reInit webApi with webview and nodes if failed (#1262)
* reInit webApi with webview and nodes if failed
if user closes and reopens the app

* after code review changes made
organized imports

* fixed melos format-check failure
