+++ app/lib/page/page.dart
@@ -1,0 +1,5 @@
+///TODO(Azamat): Add other exports here
+export 'assets/index.dart';
+export 'profile/contacts/contacts_page.dart';
+export 'profile/index.dart';
+export 'qr_scan/qr_scan_page.dart';
