+++ app/lib/service/service.dart
@@ -1,0 +1,7 @@
+///TODO(Azamat): Add other exports here
+export 'init_web_api/init_web_api.dart';
+export 'log/log_service.dart';
+export 'meetup/notification_handler.dart';
+export 'notification/lib/notification.dart';
+export 'substrate_api/api.dart';
+export 'deep_link/deep_link.dart';
